requirements.txt
bokeh==2.4.2
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
colorcet==3.0.0
dash==2.0.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2021.12.0
datashader==0.13.0
datashape==0.5.2
distributed==2021.12.0
Flask==2.0.2
Flask-Compress==1.10.1
fsspec==2021.11.1
gunicorn==20.1.0
HeapDict==1.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
llvmlite==0.37.0
locket==0.2.1
MarkupSafe==2.0.1
msgpack==1.0.3
multipledispatch==0.6.0
numba==0.54.1
numpy==1.20.3
packaging==21.3
pandas==1.3.5
param==1.12.0
partd==1.2.0
Pillow==8.4.0
plotly==5.4.0
psutil==5.8.0
pyct==0.4.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
scipy==1.7.3
six==1.16.0
sortedcontainers==2.4.0
tblib==1.7.0
tenacity==8.0.1
toolz==0.11.2
tornado==6.1
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
xarray==0.20.2
zict==2.0.0
